# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tribe management system API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Scout {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** Kvutza group (e.g., gizra name) */
  gizra: string;
  /** Role in tribe: marcaz_vogger, marcaz_tzair, roshatz, roshgad, chanich */
  role: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateScoutBody {
  name: string;
  /** @nullable */
  phone?: string | null;
  gizra: string;
  role: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateScoutBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  gizra?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Event {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** hadracha or logistics */
  category: string;
  /** upcoming, active, completed, cancelled */
  status: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  budgetAllocated?: number | null;
  createdAt: string;
}

export interface CreateEventBody {
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  status: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  budgetAllocated?: number | null;
}

export interface UpdateEventBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  budgetAllocated?: number | null;
}

export interface Attendance {
  id: number;
  scoutId: number;
  eventId: number;
  /** present, absent, late, excused */
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  scoutName?: string | null;
  /** @nullable */
  eventName?: string | null;
  createdAt: string;
}

export interface CreateAttendanceBody {
  scoutId: number;
  eventId: number;
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateAttendanceBody {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface BudgetItem {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  /** income or expense */
  type: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  date?: string | null;
  createdAt: string;
}

export interface CreateBudgetItemBody {
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  type: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  date?: string | null;
}

export interface UpdateBudgetItemBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  date?: string | null;
}

export interface ProcurementOrder {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  /** @nullable */
  supplier?: string | null;
  /** pending, approved, ordered, received, cancelled */
  status: string;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  requestedBy?: string | null;
  createdAt: string;
}

export interface CreateProcurementOrderBody {
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  /** @nullable */
  supplier?: string | null;
  status: string;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  requestedBy?: string | null;
}

export interface UpdateProcurementOrderBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  requestedBy?: string | null;
}

export interface RolePermission {
  id: number;
  role: string;
  /** hadracha or logistics */
  section: string;
  canAccess: boolean;
  createdAt: string;
}

export interface UpdatePermissionsBody {
  role: string;
  section: string;
  canAccess: boolean;
}

export type DashboardStatsRecentActivityItem = {
  type: string;
  message: string;
  timestamp: string;
};

export interface DashboardStats {
  totalScouts: number;
  totalEvents: number;
  upcomingEvents: number;
  totalBudget: number;
  totalExpenses: number;
  pendingOrders: number;
  recentActivity: DashboardStatsRecentActivityItem[];
}

export type BudgetSummaryByCategoryItem = {
  category: string;
  amount: number;
};

export interface BudgetSummary {
  totalIncome: number;
  totalExpenses: number;
  balance: number;
  byCategory: BudgetSummaryByCategoryItem[];
}

export interface AttendanceSummaryItem {
  eventId: number;
  eventName: string;
  totalRegistered: number;
  present: number;
  absent: number;
  late: number;
  excused: number;
}

export type ListScoutsParams = {
  /**
   * Filter by gizra (kvutza group)
   * @nullable
   */
  gizra?: string | null;
  /**
   * Filter by role
   * @nullable
   */
  role?: string | null;
};

export type ListEventsParams = {
  /**
   * Filter by category (hadracha/logistics)
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
};

export type ListAttendanceParams = {
  /**
   * @nullable
   */
  eventId?: number | null;
  /**
   * @nullable
   */
  scoutId?: number | null;
};

export type ListBudgetItemsParams = {
  /**
   * @nullable
   */
  category?: string | null;
};

export type ListProcurementOrdersParams = {
  /**
   * @nullable
   */
  status?: string | null;
};