# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tribe management system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by category (hadracha/logistics)
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- ATTENDANCE ----
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: eventId
          in: query
          schema:
            type: ["integer", "null"]
        - name: scoutId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceBody"
      responses:
        "201":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendanceBody"
      responses:
        "200":
          description: Updated attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Delete attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- BUDGET ----
  /budget:
    get:
      operationId: listBudgetItems
      tags: [budget]
      summary: List budget items
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of budget items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetItem"
    post:
      operationId: createBudgetItem
      tags: [budget]
      summary: Create a budget item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetItemBody"
      responses:
        "201":
          description: Budget item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"
  /budget/{id}:
    patch:
      operationId: updateBudgetItem
      tags: [budget]
      summary: Update a budget item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetItemBody"
      responses:
        "200":
          description: Updated budget item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"
    delete:
      operationId: deleteBudgetItem
      tags: [budget]
      summary: Delete a budget item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- PROCUREMENT ----
  /procurement:
    get:
      operationId: listProcurementOrders
      tags: [procurement]
      summary: List procurement orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of procurement orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcurementOrder"
    post:
      operationId: createProcurementOrder
      tags: [procurement]
      summary: Create a procurement order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcurementOrderBody"
      responses:
        "201":
          description: Procurement order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementOrder"
  /procurement/{id}:
    patch:
      operationId: updateProcurementOrder
      tags: [procurement]
      summary: Update a procurement order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProcurementOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementOrder"
    delete:
      operationId: deleteProcurementOrder
      tags: [procurement]
      summary: Delete a procurement order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- PERMISSIONS ----
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          description: "hadracha or logistics"
        status:
          type: string
          description: "upcoming, active, completed, cancelled"
        date:
          type: ["string", "null"]
          format: date-time
        location:
          type: ["string", "null"]
        budgetAllocated:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - status
        - createdAt
    CreateEventBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        date:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        budgetAllocated:
          type: ["number", "null"]
      required:
        - name
        - category
        - status
    UpdateEventBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        budgetAllocated:
          type: ["number", "null"]
    Attendance:
      type: object
      properties:
        id:
          type: integer
        scoutId:
          type: integer
        eventId:
          type: integer
        status:
          type: string
          description: "present, absent, late, excused"
        notes:
          type: ["string", "null"]
        scoutName:
          type: ["string", "null"]
        eventName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - scoutId
        - eventId
        - status
        - createdAt
    CreateAttendanceBody:
      type: object
      properties:
        scoutId:
          type: integer
        eventId:
          type: integer
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - scoutId
        - eventId
        - status
    UpdateAttendanceBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    BudgetItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        type:
          type: string
          description: "income or expense"
        category:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        date:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - amount
        - type
        - createdAt
    CreateBudgetItemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        type:
          type: string
        category:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        date:
          type: ["string", "null"]
      required:
        - title
        - amount
        - type
    UpdateBudgetItemBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        type:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        date:
          type: ["string", "null"]
    ProcurementOrder:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        supplier:
          type: ["string", "null"]
        status:
          type: string
          description: "pending, approved, ordered, received, cancelled"
        eventId:
          type: ["integer", "null"]
        requestedBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - amount
        - status
        - createdAt
    CreateProcurementOrderBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        supplier:
          type: ["string", "null"]
        status:
          type: string
        eventId:
          type: ["integer", "null"]
        requestedBy:
          type: ["string", "null"]
      required:
        - title
        - amount
        - status
    UpdateProcurementOrderBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        supplier:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        requestedBy:
          type: ["string", "null"]
...